moving arm template parsing and fixing linting
