+++ pkg/rad/armtemplate/testdata/frontend-backend.json
